Server: catch up on missed ticks (#218)
* server: warn on missed ticks

On heavy load, or when upgrading the server, the plugin might not
tick every seconds. To better quantify this, log when some ticks
are missed.

* server: catch up on missed ticks

Catches up ticks up to 10 minutes (as longer may overload the server
when starting up after being stopped a very long time).

If it takes longer to catch up for ticks that a tick, the missed
ticks will be caught up on the next turn.